bug fix in codes-workload-dump and dumpi trace workload method

+++ codes/codes-workload.h
@@ -187,7 +187,7 @@ struct codes_workload_op
             int data_type; /* MPI data type to be matched with the recv */
             int count; /* number of elements to be received */
             int tag; /* tag of the message */
-            int16_t req_id;
+            int32_t req_id;
         } send;
         struct {
             /* TODO: not sure why source rank is here */
@@ -197,7 +197,7 @@ struct codes_workload_op
             int data_type; /* MPI data type to be matched with the send */
             int count; /* number of elements to be sent */
             int tag; /* tag of the message */
-            int16_t req_id;
+            int32_t req_id;
         } recv;
         /* TODO: non-stub for other collectives */
         struct {
@@ -205,10 +205,10 @@ struct codes_workload_op
         } collective;
         struct {
             int count;
-            int16_t* req_ids;
+            int32_t* req_ids;
         } waits;
         struct {
-            int16_t req_id;
+            int32_t req_id;
         } wait;
     }u;
 };
